requirements.txt
line-bot-sdk==2.3.0
flask==2.2.2
requests==2.28.1
openai==0.25.0